pyproject.toml
[project]
name = "shl-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.9.0",
    "fastapi==0.115.0",
    "google-genai>=1.0.0",
    "numpy>=1.26.0",
    "pydantic==2.9.0",
    "uvicorn==0.30.6",
]
